OpenBR v0.4.1 (new formula)
OpenBR is an open source framework, supplying face recognition and related functionality.  OpenBR is a a communal biometrics framework supporting the development of open algorithms and reproducible evaluations. OpenBR is supported on Windows, Mac OS X, and Debian Linux. The project is licensed under Apache 2.0 and releases follow the Semantic Versioning convention. Internally the code base uses the CMake build system and requires Qt and OpenCV.